refactor: split EditViewModel responsibilities into focused implementation classes
